Fixed link for WAPM documentation and a few nits
* The "WAPM documentation" link now points correctly to https://docs.wasmer.io/ecosystem/wapm (was previously a dead link).
* Changed casing for a few words to be more correct or consistent.

One issue: the [wapm.io guide page](https://wapm.io/help/guide) seems to differ from the [WAPM page on the wasmer.io documentation website](https://docs.wasmer.io/ecosystem/wapm): the formers uses "wasm", the latter "WASM". For example, the "What is wapm?" section header should probably read "What is WAPM?" to make it more consistent with the other page. (That, or the reverse can be done, I suppose.) Likewise with the other instances of "wasm" on that page. Of course, if the command-line program name is being referred to, then it should probably be formatted as `wasm`. I'd be happy to submit a PR for the wapm.io guide, but I'm not sure that has a repo anywhere...